backend/pyproject.toml
[project]
name = "board-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2",
    "django-model-utils>=5.0.0",
    "djangorestframework>=3.16.0",
    "python-lorem-ipsum>=2.0.0",
    "rules>=3.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
